+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Parallel mode will automatically be set in a distributed pytorch scenario for
+  all processes that aren't local/node rank 0
+
+### Fixed
+- Parallel mode causing crash if global args indices are not specified.
+
+
+
+
 ## [0.6.1] - 2022-02-14
 
 ### Changed